Remove the need for `context` in `beforeAll`
While debugging things for a customer, we found that `beforeAll` doesn't
give us a context that we can use. While running the local tests, I
found that I could initialize the takeDOMSnapshot function lazily when
needed. This will simplify setup for users of this library.

I'm making this a "soft" breaking change by generating a warning log if
you pass a context to happoPlaywright.init. This way people can discover
the simplification without much disruption.