Simplify usage of library
This is a breaking change. Instead of having everyone add their own
beforeAll/afterAll hooks, we can do it for them. This makes the only
export from the library `screenshot`. As long as you use `screenshot`,
the relevant hooks will be set up for you.

This will also fix an issue where the happo-e2e library can't
monkey-patch certain browser APIs. By using a beforeEach, we can ensure
that the browser bundle is always injected as an init script.

The app2.spec.ts file was added to make sure that:
- beforeEach/beforeAll/afterAll is always called across multiple spec
  files
- we can use a custom fixture and it will still work with our own hooks
  set up in happo-playwright